pyproject.toml
[project]
name = "ragproductionapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "google-generativeai>=0.8.5",
    "inngest>=0.5.10",
    "llama-index-core>=0.14.6",
    "llama-index-readers-file>=0.5.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
